requirements.txt
Django>=5.0.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
python-decouple>=3.8 